Implemented correlation function
This commit also includes a function to retrieve a dataframe from the file, as well as a function to isolate columns from that dataframe.

After some experimentation, I discovered that they can be rendered moot with a single line of code in any other function.

Despite this, I decided to leave them be for future reference.