fix: resolved sampling issues
This bugfix also contains a new graphing function, a bar chart. Managed to draw from an equal sample of students to ensure objectivity of data in the chart.

Note to self: Consider replacing "count" index for chart with percentages if the current chart is proven to be visually confusing or unhelpful.